# docs/design/generated/OPENAPI.gen.json
{
  "components": {
    "schemas": {
...
      "Body_upload_image": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_image",
        "type": "object"
      },
...
      "CreateDocument": {
        "additionalProperties": false,
        "properties": {
          "department_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Department Id",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "department_id"
        ],
        "title": "CreateDocument",
        "type": "object"
      },
...
      "DocumentsRow": {
        "description": "documentsのDDL由来の行型。",
        "properties": {
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "department_id": {
            "title": "Department Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version Id"
          },
          "next_version": {
            "title": "Next Version",
            "type": "integer"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "shared_departments": {
            "title": "Shared Departments",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "department_id",
          "title",
          "created_by",
          "visibility",
          "shared_departments",
          "status",
          "revision",
          "next_version",
          "latest_version_id",
          "updated_at"
        ],
        "title": "DocumentsRow",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "KotoRelay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "operationId": "list_documents",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "read",
              "enum": [
                "read",
                "work",
                "manage"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "department_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "enum": [
                "",
                "active",
                "withdrawn",
                "deleted"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": false,
              "title": "Page",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "items": {
                        "$ref": "#/components/schemas/DocumentsRow"
                      },
                      "type": "array"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  ],
                  "title": "Response List Documents"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "閲覧可能な文書を検索",
        "tags": [
          "文書"
        ],
        "x-api-package": "operations/documents/list_documents",
        "x-auth-mode": "bearer",
        "x-test-samples": [
          {
            "expected_fields": {
              "code": "unauthenticated"
            },
            "expected_status": 401,
            "method": "GET",
            "name": "認証情報がない要求の拒否",
            "path": "/api/documents"
          }
        ]
      },
      "post": {
        "operationId": "create_document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "文書を作成",
        "tags": [
          "文書"
        ],
        "x-api-package": "operations/documents/create_document",
        "x-auth-mode": "bearer",
        "x-test-samples": [
          {
            "expected_fields": {
              "code": "unauthenticated"
            },
            "expected_status": 401,
            "method": "POST",
            "name": "認証情報がない要求の拒否",
            "path": "/api/documents"
          }
        ]
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "operationId": "read_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Document",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "承認版または担当版を表示",
        "tags": [
          "文書"
        ],
        "x-api-package": "operations/documents/read_document",
        "x-auth-mode": "bearer",
        "x-test-samples": [
          {
            "expected_fields": {
              "code": "unauthenticated"
            },
            "expected_status": 401,
            "method": "GET",
            "name": "認証情報がない要求の拒否",
            "path": "/api/documents/00000000-0000-0000-0000-000000000001"
          }
        ]
      }
    },
...
    "/api/images/documents/{document_id}": {
      "post": {
        "operationId": "upload_image",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Image",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "画像を添付して位置付きOCRを実行",
        "tags": [
          "画像・OCR"
        ],
        "x-api-package": "operations/images/upload_image",
        "x-auth-mode": "bearer",
        "x-test-samples": [
          {
            "expected_fields": {
              "code": "unauthenticated"
            },
            "expected_status": 401,
            "method": "POST",
            "name": "認証情報がない要求の拒否",
            "path": "/api/images/documents/00000000-0000-0000-0000-000000000001"
          }
        ]
      }
    },
...
  }
}